Extract Guard interface and create jwt_guard
